fixed search by multiple tags. solved some issues

+++ repository/src/main/java/com/epam/esm/dao/RowMapCertificateProvider.java
@@ -22,7 +22,7 @@ public class RowMapCertificateProvider implements RowMapper<Certificate> {
             certificate.setId(resultSet.getLong(PROVIDER_ID));
             certificate.setName(resultSet.getString(PROVIDER_NAME));
             certificate.setDescription(resultSet.getString(PROVIDER_DESCRIPTION));
-            certificate.setPrice(resultSet.getDouble(PROVIDER_PRICE));
+            certificate.setPrice(resultSet.getBigDecimal(PROVIDER_PRICE));
             certificate.setDuration(resultSet.getInt(PROVIDER_DURATION));
             certificate.setCreateDate(resultSet.getTimestamp(PROVIDER_CREATE_DATE) == null ? null : resultSet.getTimestamp(PROVIDER_CREATE_DATE).toInstant());
             certificate.setLastUpdateDate(resultSet.getTimestamp(PROVIDER_LAST_UPDATE_DATE) == null ? null : resultSet.getTimestamp(PROVIDER_LAST_UPDATE_DATE).toInstant());
